Make program.populateSentRegistrationEmailAddresses idempotent
So it can be rerun when e.g. a classification is deleted.

A slight change to the previous version is that updates to buyer and author email addresses are handled. Although this method is called too late in the registration process to show the updated addresses in the classification UI.