Fix recent-classification row
Turns out, is was showing the latest classification for given program. Which is NOT necessarily the one created by the user.

Changing the row to show the classification created by the user caused a hole whereby deleting that classification tried to refresh the row which shouldn't exist anymore. Fixing by reloading all recents when a row is encountered which doesn't have a classification by the author anymore.