Check that server hasn't been upgraded on each ajax request
If so, force the client to refresh the page.

This works with a npm postinstall hook, which writes the msSinceEpoch date into build.revision on heroku.
App checks that client has a 'build.revision'- cookie which matches the build.revision file content.
Cookie is set when an entrypoint html page is loaded (e.g. public.html).

(Turned out that accessing e.g. the current git sha was impossible in heroku)