Repopulate program.sentRegistrationEmailAddresses when reclassification begins
If buyer or author email addresses have changed, this shows the correct new addresses in the UI when reclassifying.

Also, since programs coming in from the XML api don't have sentRegistrationEmailAddresses populated, this fills buyer + author emails for those as well.

Repopulating sentRegistrationEmailAddresses on register is still necessary to add the authors own email address to the array. (Could be simpler tho)